feat: add github_release strategy
